+++ src/lib.rs
@@ -51,6 +51,9 @@
 //! * **`fields(Option=!)`** <br/>
 //! Disable automatic [`Option`] detection for **fields**. See ["`Option` fields"](#option-fields).
 //!
+//! * **`fields(Default=!)`** <br/>
+//! Disable automatic [`Default`] detection for **fields**. See ["`Default` fields"](#default-fields).
+//!
 //! * **`fields(Into=!)`** <br/>
 //! Disable [`Into`] for **fields**. See ["`Into` argument"](#into-argument).
 //!